requirements.txt
django
mysqlclient
bcrypt
pyjwt